requirements.txt
ansible-core
pylint==2.15.5
yamllint==1.28.0
pytest==7.2.0
pytest-forked==1.4.0
pytest-xdist==3.0.2
antsibull-docs==1.7.3
python-dotenv[cli]==0.21.0
boto3==1.26.5
awscli==1.27.5
google-auth==2.14.1
google-cloud==0.34.0
black==22.10.0
mypy==0.991

# this gets installed as part of the collection requirements to avoid dependency conflicts
# the readme installs this without dependencies to prevent conflicts
# azure-cli==2.34.0
